+++ tests/vim/test_vimwiki_convert.py
@@ -9,6 +9,7 @@ def test_vim_vader_all():
 
     Pytest monitors python coverage and Vader runs Vim python modules--i.e., we
     don't have to mock `import vim`. `run.sh` runs without Pytest.
+    Raise CalledProcessError if vim exits non-zero
     """
     vim_command = shlex.split(
         "vim -Nu tests/vim/vimrc -Es -c 'Vader! tests/vim/*.vader'"
@@ -17,7 +18,9 @@ def test_vim_vader_all():
         vim_command,
         capture_output=True,
         encoding="utf8",
-        check=True,
+        check=False,
     )
     print(test_input.stdout)
     print(test_input.stderr)
+
+    test_input.check_returncode()